added functions for summary statistics
this code adds columns for total occurences of the word for that day, total occurennces within a topic for that day, sum of all word counts for a day, and sum of all word counts in a topic for a day. These were all used to calculated the sub_freq to all_freq ratio.

Also added a save to csv, but it failed becaue of memory.