Removed Sort, added topic partition to window, made date column to get rid of window
The word explode function was a bottleneck. Removing the sort fixed it.

The remove duplicates was also a bottleneck. I moved up the remove nulls. Then added the remove duplicates. This seemed to fix it.

I added a partition for topics into the window lag function. This seemed to fix that bottleneck.

I removed the date-window column at the end and replaced it with a date column (no time shown) hopefully but testing in parquet now.

Added save as parquet to the end with topic and date as keys.